Separate StationList.vue
