+++ src/components/station/StationInfo.vue
@@ -17,43 +17,7 @@
       </v-toolbar>
       <v-divider />
       <v-card flat class="text-center" v-show="station === null">
-        <v-list>
-          <v-hover v-slot="{ isHovering, props }">
-            <v-list-item
-              v-for="(s, i) in stations"
-              :key="i"
-              v-bind="props"
-              :class="{ 'on-hover': isHovering }"
-              @click="onClick(s)"
-              @mouseover="onHover(s)"
-            >
-              <h4 class="text-left" v-html="$root.clean(s.properties.name)" />
-              <template v-slot:append>
-                <v-btn
-                  variant="outlined"
-                  size="small"
-                  color="#014e9e"
-                  @click.stop="openData(s)"
-                  class="mx-3"
-                >
-                  {{ $t("navigation.data") }}
-                  <v-icon end icon="mdi-chart-scatter-plot"></v-icon>
-                </v-btn>
-                <v-btn
-                  variant="outlined"
-                  size="small"
-                  color="#014e9e"
-                  :target="s.id"
-                  :title="s.id"
-                  :href="s.properties.url"
-                >
-                  OSCAR
-                  <v-icon end icon="mdi-open-in-new"></v-icon>
-                </v-btn>
-              </template>
-            </v-list-item>
-          </v-hover>
-        </v-list>
+        <station-list :features="features" :map="map" />
       </v-card>
 
       <v-card flat v-show="station !== null">
@@ -104,8 +68,10 @@ import "leaflet/dist/leaflet.css";
 import Plotly from "plotly.js-dist-min";
 
 import { defineComponent } from "vue";
+import StationList from "./StationList.vue";
 
 export default defineComponent({
+  components: { StationList },
   name: "StationInfo",
   template: "#station-info",
   props: ["features", "map"],
@@ -130,24 +96,6 @@ export default defineComponent({
         return this.station.properties.name;
       } else {
         return this.station;
-      }
-    },
-    stations: function () {
-      if (this.features.stations === null) {
-        return [];
-      } else {
-        const stns = [...this.features.stations.features].sort((a, b) => {
-          const nameA = a.properties.name.toUpperCase(); // ignore upper and lowercase
-          const nameB = b.properties.name.toUpperCase(); // ignore upper and lowercase
-          if (nameA < nameB) {
-            return -1;
-          } else if (nameA > nameB) {
-            return 1;
-          } else {
-            return 0;
-          }
-        });
-        return stns;
       }
     },
   },
@@ -279,25 +227,6 @@ export default defineComponent({
 
         Plotly.newPlot("stationStatus", data, layout, config);
       });
-    },
-    onClick(station) {
-      this.features_.station = station;
-      const latlng = [
-        station.geometry.coordinates[1],
-        station.geometry.coordinates[0],
-      ];
-      this.map.flyTo(latlng);
-    },
-    onHover(station) {
-      const latlng = [
-        station.geometry.coordinates[1],
-        station.geometry.coordinates[0],
-      ];
-      this.map.openPopup(station.properties.name, latlng);
-    },
-    openData(station) {
-      this.features_.station = station;
-      this.$root.toggleDialog();
     },
   },
 });
